Delete ban API v1
