--- app/Http/Requests/Api/GameBanCheckRequest.php
@@ -1,42 +1,0 @@
-<?php
-
-namespace App\Http\Requests\Api;
-
-use App\Exceptions\Http\BadRequestException;
-use Illuminate\Contracts\Validation\Validator;
-use Illuminate\Foundation\Http\FormRequest;
-
-final class GameBanCheckRequest extends FormRequest
-{
-    /**
-     * Get the validation rules that apply to the request.
-     */
-    public function rules(): array
-    {
-        return [
-            'player_id_type' => 'required',
-            'player_id' => 'required',
-        ];
-    }
-
-    /**
-     * Determine if the user is authorized to make this request.
-     */
-    public function authorize(): bool
-    {
-        return true;
-    }
-
-    /**
-     * Handle a failed validation attempt.
-     *
-     *
-     * @return void
-     *
-     * @throws \Illuminate\Validation\ValidationException
-     */
-    protected function failedValidation(Validator $validator)
-    {
-        throw new BadRequestException('bad_input', $validator->errors()->first());
-    }
-}
