--- app/Http/Requests/Api/GameUnbanRequest.php
@@ -1,44 +1,0 @@
-<?php
-
-namespace App\Http\Requests\Api;
-
-use App\Exceptions\Http\BadRequestException;
-use Illuminate\Contracts\Validation\Validator;
-use Illuminate\Foundation\Http\FormRequest;
-
-final class GameUnbanRequest extends FormRequest
-{
-    /**
-     * Get the validation rules that apply to the request.
-     */
-    public function rules(): array
-    {
-        return [
-            'player_id_type' => 'required',
-            'player_id' => 'required',
-            'banner_id_type' => 'required',
-            'banner_id' => 'required',
-        ];
-    }
-
-    /**
-     * Determine if the user is authorized to make this request.
-     */
-    public function authorize(): bool
-    {
-        return true;
-    }
-
-    /**
-     * Handle a failed validation attempt.
-     *
-     *
-     * @return void
-     *
-     * @throws \Illuminate\Validation\ValidationException
-     */
-    protected function failedValidation(Validator $validator)
-    {
-        throw new BadRequestException('bad_input', $validator->errors()->first());
-    }
-}
