--- app/Services/PlayerBans/Exceptions/MalformedTokenException.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace App\Services\PlayerBans\Exceptions;
-
-use App\Exceptions\Http\BadRequestException;
-
-/**
- * @deprecated
- */
-final class MalformedTokenException extends BadRequestException
-{
-}
