--- app/Services/PlayerBans/Exceptions/UnauthorisedKeyActionException.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace App\Services\PlayerBans\Exceptions;
-
-use App\Exceptions\Http\UnauthorisedException;
-
-/**
- * @deprecated
- */
-final class UnauthorisedKeyActionException extends UnauthorisedException
-{
-}
