--- app/Services/PlayerBans/Exceptions/UnauthorisedTokenException.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace App\Services\PlayerBans\Exceptions;
-
-use App\Exceptions\Http\UnauthorisedException;
-
-/**
- * @deprecated
- */
-final class UnauthorisedTokenException extends UnauthorisedException
-{
-}
