+++ repositories/MinecraftPlayerRepository.php
@@ -3,26 +3,30 @@
 namespace Repositories;
 
 use Entities\Models\Eloquent\MinecraftPlayer;
+use Entities\Models\MinecraftUUID;
 
 /**
  * @final
- *
- * @deprecated
  */
 class MinecraftPlayerRepository
 {
     public function store(
-        string $uuid,
+        MinecraftUUID $uuid,
         ?int $accountId = null,
     ): MinecraftPlayer {
         return MinecraftPlayer::create([
-            'uuid' => $uuid,
+            'uuid' => $uuid->rawValue(),
             'account_id' => $accountId,
         ]);
     }
 
-    public function getByUUID(string $uuid): ?MinecraftPlayer
+    public function getByUUID(MinecraftUUID $uuid): ?MinecraftPlayer
     {
-        return MinecraftPlayer::where('uuid', $uuid)->first();
+        return MinecraftPlayer::where('uuid', $uuid->rawValue())->first();
+    }
+
+    public function getById(int $id): ?MinecraftPlayer
+    {
+        return MinecraftPlayer::where('player_minecraft_id', $id)->first();
     }
 }
